# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.City:
    properties:
      alias:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      province:
        $ref: '#/definitions/models.Province'
      province_id:
        type: integer
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Country:
    properties:
      alias:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      provinces:
        items:
          $ref: '#/definitions/models.Province'
        type: array
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.Hotel:
    properties:
      address:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      email_address:
        type: string
      extra_data:
        type: string
      fax_number:
        type: string
      hotel_grade:
        $ref: '#/definitions/models.HotelGrade'
      hotel_grade_id:
        type: integer
      hotel_type:
        $ref: '#/definitions/models.HotelType'
      hotel_type_id:
        type: integer
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: integer
      phone_number1:
        type: string
      phone_number2:
        type: string
      postal_code:
        type: string
      province:
        $ref: '#/definitions/models.Province'
      province_id:
        type: integer
      tenant_id:
        type: integer
      thumbnails:
        items:
          $ref: '#/definitions/os.File'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
      website:
        type: string
    type: object
  models.HotelGrade:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      hotel_type:
        $ref: '#/definitions/models.HotelType'
      hotel_type_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.HotelType:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      grades:
        items:
          $ref: '#/definitions/models.HotelGrade'
        type: array
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.Province:
    properties:
      alias:
        type: string
      cities:
        items:
          $ref: '#/definitions/models.City'
        type: array
      country:
        $ref: '#/definitions/models.Country'
      country_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.User:
    properties:
      address:
        type: string
      birth_date:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
    type: object
  os.File:
    type: object
info:
  contact: {}
  description: Swagger documentation for reservation API .
  license:
    name: MIT
  title: Reservation API
  version: "1.0"
paths:
...
  /hotel-grades:
    delete:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.HotelGrade'
            type: array
      summary: Delete HotelGrade
      tags:
      - HotelGrade
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.HotelGrade'
            type: array
      summary: findAll HotelGrades
      tags:
      - HotelGrade
    post:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: HotelGrade
        in: body
        name: HotelGrade
        required: true
        schema:
          $ref: '#/definitions/models.HotelGrade'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HotelGrade'
      summary: create new HotelGrade
      tags:
      - HotelGrade
  /hotel-grades/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HotelGrade'
      summary: find HotelGrade by id
      tags:
      - HotelGrade
    put:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      - description: HotelGrade
        in: body
        name: HotelGrade
        required: true
        schema:
          $ref: '#/definitions/models.HotelGrade'
      - description: HotelGrade
        in: body
        name: HotelGrade
        required: true
        schema:
          $ref: '#/definitions/models.HotelGrade'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HotelGrade'
      summary: update HotelGrade
      tags:
      - HotelGrade
...
  /hotels:
    delete:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Hotel'
            type: array
      summary: Delete Hotel
      tags:
      - Hotel
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Hotel'
            type: array
      summary: findAll Hotels
      tags:
      - Hotel
    post:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Hotel
        in: body
        name: Hotel
        required: true
        schema:
          $ref: '#/definitions/models.Hotel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hotel'
      summary: create new Hotel
      tags:
      - Hotel
  /hotels/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hotel'
      summary: find Hotel by id
      tags:
      - Hotel
    put:
      consumes:
      - application/json
      parameters:
      - description: X-Tenant-ID
        in: header
        name: X-Tenant-ID
        required: true
        type: integer
      - description: Id
        in: path
        name: Id
        required: true
        type: integer
      - description: Hotel
        in: body
        name: Hotel
        required: true
        schema:
          $ref: '#/definitions/models.Hotel'
      - description: Hotel
        in: body
        name: Hotel
        required: true
        schema:
          $ref: '#/definitions/models.Hotel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hotel'
      summary: update Hotel
      tags:
      - Hotel
...
swagger: "2.0"